+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic parsers: `any`, `str`, `re`, `or`, `seq`, `many`, `many1`, `maybe`,
   `not`, `tag`
+- `fail` creates a labelled failure, which is like a throw for parsers
 - `cstToAst` transforms a CST to an AST
 
 [unreleased]: https://github.com/unleashy/vahv/compare/v0.1.0...HEAD